+++ backend/src/Controller/BidOrder/BidOrderController.php
@@ -6,8 +6,7 @@ use App\AutoMapping;
 use App\Constant\Main\MainErrorConstant;
 use App\Constant\StoreOwner\StoreProfileConstant;
 use App\Controller\BaseController;
-use App\Request\BidOrder\BidOrderCreateRequest;
-use App\Request\BidOrder\BidOrderFilterBySupplierRequest;
+use App\Request\Order\BidOrderFilterBySupplierRequest;
 use App\Service\BidOrder\BidOrderService;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Nelmio\ApiDocBundle\Annotation\Security;
@@ -118,58 +117,58 @@ class BidOrderController extends BaseController
 //        return $this->response($result, self::CREATE);
 //    }
 
-    /**
-     * supplier: filter bid orders which the supplier had not provide a price offer for any one of them yet.
-     * @Route("filterbidordersbysupplier", name="filterBidOrdersBySupplier", methods={"POST"})
-     * @IsGranted("ROLE_SUPPLIER")
-     * @param Request $request
-     * @return JsonResponse
-     *
-     * @OA\Tag(name="Bid Order")
-     *
-     * @OA\Parameter(
-     *      name="token",
-     *      in="header",
-     *      description="token to be passed as a header",
-     *      required=true
-     * )
-     *
-     * @OA\RequestBody(
-     *      description="filtering options",
-     *      @OA\JsonContent(
-     *          @OA\Property(type="string", property="fromDate"),
-     *          @OA\Property(type="string", property="toDate")
-     *      )
-     * )
-     *
-     * @OA\Response(
-     *      response=201,
-     *      description="Returns the bid orders info",
-     *      @OA\JsonContent(
-     *          @OA\Property(type="string", property="status_code"),
-     *          @OA\Property(type="string", property="msg"),
-     *          @OA\Property(type="array", property="Data",
-     *              @OA\Items(
-     *                  ref=@Model(type="App\Response\BidOrder\BidOrderFilterBySupplierResponse")
-     *              )
-     *      )
-     *   )
-     * )
-     *
-     * @Security(name="Bearer")
-     */
-    public function filterBidOrdersBySupplier(Request $request): JsonResponse
-    {
-        $data = json_decode($request->getContent(), true);
-
-        $request = $this->autoMapping->map(stdClass::class, BidOrderFilterBySupplierRequest::class, (object)$data);
-
-        $request->setSupplierId($this->getUserId());
-
-        $result = $this->bidOrderService->filterBidOrdersBySupplier($request);
-
-        return $this->response($result, self::FETCH);
-    }
+//    /**
+//     * supplier: filter bid orders which the supplier had not provide a price offer for any one of them yet.
+//     * @Route("filterbidordersbysupplier", name="filterBidOrdersBySupplier", methods={"POST"})
+//     * @IsGranted("ROLE_SUPPLIER")
+//     * @param Request $request
+//     * @return JsonResponse
+//     *
+//     * @OA\Tag(name="Bid Order")
+//     *
+//     * @OA\Parameter(
+//     *      name="token",
+//     *      in="header",
+//     *      description="token to be passed as a header",
+//     *      required=true
+//     * )
+//     *
+//     * @OA\RequestBody(
+//     *      description="filtering options",
+//     *      @OA\JsonContent(
+//     *          @OA\Property(type="string", property="fromDate"),
+//     *          @OA\Property(type="string", property="toDate")
+//     *      )
+//     * )
+//     *
+//     * @OA\Response(
+//     *      response=201,
+//     *      description="Returns the bid orders info",
+//     *      @OA\JsonContent(
+//     *          @OA\Property(type="string", property="status_code"),
+//     *          @OA\Property(type="string", property="msg"),
+//     *          @OA\Property(type="array", property="Data",
+//     *              @OA\Items(
+//     *                  ref=@Model(type="App\Response\Order\BidOrderFilterBySupplierResponse")
+//     *              )
+//     *      )
+//     *   )
+//     * )
+//     *
+//     * @Security(name="Bearer")
+//     */
+//    public function filterBidOrdersBySupplier(Request $request): JsonResponse
+//    {
+//        $data = json_decode($request->getContent(), true);
+//
+//        $request = $this->autoMapping->map(stdClass::class, BidOrderFilterBySupplierRequest::class, (object)$data);
+//
+//        $request->setSupplierId($this->getUserId());
+//
+//        $result = $this->bidOrderService->filterBidOrdersBySupplier($request);
+//
+//        return $this->response($result, self::FETCH);
+//    }
 
     /**
      * supplier: get bid order details by id for supplier
@@ -208,58 +207,58 @@ class BidOrderController extends BaseController
         return $this->response($result, self::FETCH);
     }
 
-    /**
-     * supplier: filter bid orders which the supplier had provided a price offer for each one of them.
-     * @Route("filterbidorderswithpriceoffersbysupplier", name="filterBidOrdersWithPriceOffersBySupplier", methods={"POST"})
-     * @IsGranted("ROLE_SUPPLIER")
-     * @param Request $request
-     * @return JsonResponse
-     *
-     * @OA\Tag(name="Bid Order")
-     *
-     * @OA\Parameter(
-     *      name="token",
-     *      in="header",
-     *      description="token to be passed as a header",
-     *      required=true
-     * )
-     *
-     * @OA\RequestBody(
-     *      description="filtering options",
-     *      @OA\JsonContent(
-     *          @OA\Property(type="string", property="fromDate"),
-     *          @OA\Property(type="string", property="toDate"),
-     *          @OA\Property(type="integer", property="priceOfferStatus"),
-     *          @OA\Property(type="boolean", property="openToPriceOffer")
-     *      )
-     * )
-     *
-     * @OA\Response(
-     *      response=201,
-     *      description="Returns the bid orders info",
-     *      @OA\JsonContent(
-     *          @OA\Property(type="string", property="status_code"),
-     *          @OA\Property(type="string", property="msg"),
-     *          @OA\Property(type="array", property="Data",
-     *              @OA\Items(
-     *                  ref=@Model(type="App\Response\BidOrder\BidOrderFilterBySupplierResponse")
-     *              )
-     *      )
-     *   )
-     * )
-     *
-     * @Security(name="Bearer")
-     */
-    public function filterBidOrdersThatHavePriceOffersBySupplier(Request $request): JsonResponse
-    {
-        $data = json_decode($request->getContent(), true);
-
-        $request = $this->autoMapping->map(stdClass::class, BidOrderFilterBySupplierRequest::class, (object)$data);
-
-        $request->setSupplierId($this->getUserId());
-
-        $result = $this->bidOrderService->filterBidOrdersThatHavePriceOffersBySupplier($request);
-
-        return $this->response($result, self::FETCH);
-    }
+//    /**
+//     * supplier: filter bid orders which the supplier had provided a price offer for each one of them.
+//     * @Route("filterbidorderswithpriceoffersbysupplier", name="filterBidOrdersWithPriceOffersBySupplier", methods={"POST"})
+//     * @IsGranted("ROLE_SUPPLIER")
+//     * @param Request $request
+//     * @return JsonResponse
+//     *
+//     * @OA\Tag(name="Bid Order")
+//     *
+//     * @OA\Parameter(
+//     *      name="token",
+//     *      in="header",
+//     *      description="token to be passed as a header",
+//     *      required=true
+//     * )
+//     *
+//     * @OA\RequestBody(
+//     *      description="filtering options",
+//     *      @OA\JsonContent(
+//     *          @OA\Property(type="string", property="fromDate"),
+//     *          @OA\Property(type="string", property="toDate"),
+//     *          @OA\Property(type="integer", property="priceOfferStatus"),
+//     *          @OA\Property(type="boolean", property="openToPriceOffer")
+//     *      )
+//     * )
+//     *
+//     * @OA\Response(
+//     *      response=201,
+//     *      description="Returns the bid orders info",
+//     *      @OA\JsonContent(
+//     *          @OA\Property(type="string", property="status_code"),
+//     *          @OA\Property(type="string", property="msg"),
+//     *          @OA\Property(type="array", property="Data",
+//     *              @OA\Items(
+//     *                  ref=@Model(type="App\Response\Order\BidOrderFilterBySupplierResponse")
+//     *              )
+//     *      )
+//     *   )
+//     * )
+//     *
+//     * @Security(name="Bearer")
+//     */
+//    public function filterBidOrdersThatHavePriceOffersBySupplier(Request $request): JsonResponse
+//    {
+//        $data = json_decode($request->getContent(), true);
+//
+//        $request = $this->autoMapping->map(stdClass::class, BidOrderFilterBySupplierRequest::class, (object)$data);
+//
+//        $request->setSupplierId($this->getUserId());
+//
+//        $result = $this->bidOrderService->filterBidOrdersThatHavePriceOffersBySupplier($request);
+//
+//        return $this->response($result, self::FETCH);
+//    }
 }
