+++ backend/src/Repository/BidOrderEntityRepository.php
@@ -7,7 +7,7 @@ use App\Entity\ImageEntity;
 use App\Entity\PriceOfferEntity;
 use App\Entity\SupplierCategoryEntity;
 use App\Entity\SupplierProfileEntity;
-use App\Request\BidOrder\BidOrderFilterBySupplierRequest;
+use App\Request\Order\BidOrderFilterBySupplierRequest;
 use DateTime;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\ORM\Query\Expr\Join;
@@ -26,62 +26,62 @@ class BidOrderEntityRepository extends ServiceEntityRepository
         parent::__construct($registry, BidOrderEntity::class);
     }
 
-    // This function filter bid orders which the supplier had not provide a price offer for any one of them yet.
-    public function filterBidOrdersBySupplier(BidOrderFilterBySupplierRequest $request): array
-    {
-        $query = $this->createQueryBuilder('bidOrderEntity')
-            ->select('bidOrderEntity.id', 'bidOrderEntity.title', 'bidOrderEntity.description', 'bidOrderEntity.createdAt', 'bidOrderEntity.updatedAt')
-
-            ->andWhere('bidOrderEntity.openToPriceOffer = :openToPriceOfferStatus')
-            ->setParameter('openToPriceOfferStatus', 1)
-
-            ->leftJoin(
-                SupplierCategoryEntity::class,
-                'supplierCategoryEntity',
-                Join::WITH,
-                'supplierCategoryEntity.id = bidOrderEntity.supplierCategory'
-            )
-
-            ->leftJoin(
-                SupplierProfileEntity::class,
-                'supplierProfileEntity',
-                Join::WITH,
-                'supplierCategoryEntity.id IN (:supCategory)'
-            )
-
-            ->setParameter('supCategory', $this->getSupplierCategoriesBySupplierId($request->getSupplierId()))
-
-            ->andWhere('supplierProfileEntity.user = :supplierId')
-            ->setParameter('supplierId', $request->getSupplierId())
-
-            ->orderBy('bidOrderEntity.id', 'DESC');
-
-        //---- Check if bid order is among the orders which the supplier had made a previous offer for it
-        $bidOrderIds = $this->getBidOrderIdsBySupplierIdAndThatHavePriceOffers($request->getSupplierId());
-        if (! empty($bidOrderIds)) {
-            $query->andWhere('bidOrderEntity.id NOT IN (:bidOrderIds)');
-            $query->setParameter('bidOrderIds', $bidOrderIds);
-        }
-        //---- End checking block
-
-        if (($request->getFromDate() != null || $request->getFromDate() != "") && ($request->getToDate() === null || $request->getToDate() === "")) {
-            $query->andWhere('bidOrderEntity.createdAt >= :createdAt');
-            $query->setParameter('createdAt', $request->getFromDate());
-
-        } elseif (($request->getFromDate() === null || $request->getFromDate() === "") && ($request->getToDate() != null || $request->getToDate() != "")) {
-            $query->andWhere('bidOrderEntity.createdAt <= :createdAt');
-            $query->setParameter('createdAt', (new DateTime($request->getToDate()))->modify('+1 day')->format('Y-m-d'));
-
-        } elseif (($request->getFromDate() != null || $request->getFromDate() != "") && ($request->getToDate() != null || $request->getToDate() != "")) {
-            $query->andWhere('bidOrderEntity.createdAt >= :fromDate');
-            $query->setParameter('fromDate', $request->getFromDate());
-
-            $query->andWhere('bidOrderEntity.createdAt <= :toDate');
-            $query->setParameter('toDate', (new DateTime($request->getToDate()))->modify('+1 day')->format('Y-m-d'));
-        }
-
-        return $query->getQuery()->getResult();
-    }
+//    // This function filter bid orders which the supplier had not provide a price offer for any one of them yet.
+//    public function filterBidOrdersBySupplier(BidOrderFilterBySupplierRequest $request): array
+//    {
+//        $query = $this->createQueryBuilder('bidOrderEntity')
+//            ->select('bidOrderEntity.id', 'bidOrderEntity.title', 'bidOrderEntity.description', 'bidOrderEntity.createdAt', 'bidOrderEntity.updatedAt')
+//
+//            ->andWhere('bidOrderEntity.openToPriceOffer = :openToPriceOfferStatus')
+//            ->setParameter('openToPriceOfferStatus', 1)
+//
+//            ->leftJoin(
+//                SupplierCategoryEntity::class,
+//                'supplierCategoryEntity',
+//                Join::WITH,
+//                'supplierCategoryEntity.id = bidOrderEntity.supplierCategory'
+//            )
+//
+//            ->leftJoin(
+//                SupplierProfileEntity::class,
+//                'supplierProfileEntity',
+//                Join::WITH,
+//                'supplierCategoryEntity.id IN (:supCategory)'
+//            )
+//
+//            ->setParameter('supCategory', $this->getSupplierCategoriesBySupplierId($request->getSupplierId()))
+//
+//            ->andWhere('supplierProfileEntity.user = :supplierId')
+//            ->setParameter('supplierId', $request->getSupplierId())
+//
+//            ->orderBy('bidOrderEntity.id', 'DESC');
+//
+//        //---- Check if bid order is among the orders which the supplier had made a previous offer for it
+//        $bidOrderIds = $this->getBidOrderIdsBySupplierIdAndThatHavePriceOffers($request->getSupplierId());
+//        if (! empty($bidOrderIds)) {
+//            $query->andWhere('bidOrderEntity.id NOT IN (:bidOrderIds)');
+//            $query->setParameter('bidOrderIds', $bidOrderIds);
+//        }
+//        //---- End checking block
+//
+//        if (($request->getFromDate() != null || $request->getFromDate() != "") && ($request->getToDate() === null || $request->getToDate() === "")) {
+//            $query->andWhere('bidOrderEntity.createdAt >= :createdAt');
+//            $query->setParameter('createdAt', $request->getFromDate());
+//
+//        } elseif (($request->getFromDate() === null || $request->getFromDate() === "") && ($request->getToDate() != null || $request->getToDate() != "")) {
+//            $query->andWhere('bidOrderEntity.createdAt <= :createdAt');
+//            $query->setParameter('createdAt', (new DateTime($request->getToDate()))->modify('+1 day')->format('Y-m-d'));
+//
+//        } elseif (($request->getFromDate() != null || $request->getFromDate() != "") && ($request->getToDate() != null || $request->getToDate() != "")) {
+//            $query->andWhere('bidOrderEntity.createdAt >= :fromDate');
+//            $query->setParameter('fromDate', $request->getFromDate());
+//
+//            $query->andWhere('bidOrderEntity.createdAt <= :toDate');
+//            $query->setParameter('toDate', (new DateTime($request->getToDate()))->modify('+1 day')->format('Y-m-d'));
+//        }
+//
+//        return $query->getQuery()->getResult();
+//    }
 
     public function getBidOrderByIdForSupplier(int $bidOrderId): ?BidOrderEntity
     {
@@ -94,37 +94,37 @@ class BidOrderEntityRepository extends ServiceEntityRepository
             ->getOneOrNullResult();
     }
 
-    // This function returns array of bid orders Ids that the supplier had made a price offer for them
-    public function getBidOrderIdsBySupplierIdAndThatHavePriceOffers(int $supplierId): array
-    {
-        return $this->createQueryBuilder('bidOrderEntity')
-            ->select('bidOrderEntity.id')
-
-            ->andWhere('bidOrderEntity.openToPriceOffer = :openToPriceOfferStatus')
-            ->setParameter('openToPriceOfferStatus', 1)
-
-            ->leftJoin(
-                PriceOfferEntity::class,
-                'priceOfferEntity',
-                Join::WITH,
-                'priceOfferEntity.bidOrder = bidOrderEntity.id'
-            )
-
-            ->leftJoin(
-                SupplierProfileEntity::class,
-                'supplierProfileEntity',
-                Join::WITH,
-                'supplierProfileEntity.id = priceOfferEntity.supplierProfile'
-            )
-
-            ->andWhere('supplierProfileEntity.user = :supplierId')
-            ->setParameter('supplierId', $supplierId)
-
-            ->orderBy('bidOrderEntity.id', 'DESC')
-
-            ->getQuery()
-            ->getSingleColumnResult();
-    }
+//    // This function returns array of bid orders Ids that the supplier had made a price offer for them
+//    public function getBidOrderIdsBySupplierIdAndThatHavePriceOffers(int $supplierId): array
+//    {
+//        return $this->createQueryBuilder('bidOrderEntity')
+//            ->select('bidOrderEntity.id')
+//
+//            ->andWhere('bidOrderEntity.openToPriceOffer = :openToPriceOfferStatus')
+//            ->setParameter('openToPriceOfferStatus', 1)
+//
+//            ->leftJoin(
+//                PriceOfferEntity::class,
+//                'priceOfferEntity',
+//                Join::WITH,
+//                'priceOfferEntity.bidOrder = bidOrderEntity.id'
+//            )
+//
+//            ->leftJoin(
+//                SupplierProfileEntity::class,
+//                'supplierProfileEntity',
+//                Join::WITH,
+//                'supplierProfileEntity.id = priceOfferEntity.supplierProfile'
+//            )
+//
+//            ->andWhere('supplierProfileEntity.user = :supplierId')
+//            ->setParameter('supplierId', $supplierId)
+//
+//            ->orderBy('bidOrderEntity.id', 'DESC')
+//
+//            ->getQuery()
+//            ->getSingleColumnResult();
+//    }
 
     // This function filter bid orders which have price offers made by the supplier (who request the filter).
     public function filterBidOrdersThatHavePriceOffersBySupplier(BidOrderFilterBySupplierRequest $request): array
@@ -152,19 +152,6 @@ class BidOrderEntityRepository extends ServiceEntityRepository
             ->andWhere('priceOfferEntity.supplierProfile = supplierProfileEntity.id')
 
             ->orderBy('bidOrderEntity.id', 'DESC');
-
-        if ($request->getPriceOfferStatus()) {
-//            $orders = $query->getQuery()->getResult();
-//
-//            if (! empty($orders)) {
-//                foreach ($orders as $order) {
-//                    $ordersIDs[] = $order['id'];
-//                }
-//
-//                $pricesOffersIDs = $this->getLastPriceOfferIdsArrayBySupplierIdAndPriceOfferStatus($request->getSupplierId(), $ordersIDs);
-//                dd($pricesOffersIDs);
-//            }
-        }
 
         if ($request->getOpenToPriceOffer() !== null) {
             $query->andWhere('bidOrderEntity.openToPriceOffer = :openToPriceOffer');
@@ -212,27 +199,27 @@ class BidOrderEntityRepository extends ServiceEntityRepository
             ->getSingleResult();
     }
 
-    public function getSupplierCategoriesBySupplierId(int $supplierId): array
-    {
-        $query = $this->createQueryBuilder('bidOrder')
-            ->select('DISTINCT(supplierProfileEntity.supplierCategories)')
-
-            ->leftJoin(
-                SupplierProfileEntity::class,
-                'supplierProfileEntity',
-                Join::WITH,
-                'supplierProfileEntity.user = :userId'
-            )
-
-            ->setParameter('userId', $supplierId)
-
-            ->getQuery()
-            ->getSingleColumnResult();
-
-        if (empty($query)) {
-            return $query;
-        }
-
-        return json_decode($query[0]);
-    }
+//    public function getSupplierCategoriesBySupplierId(int $supplierId): array
+//    {
+//        $query = $this->createQueryBuilder('bidOrder')
+//            ->select('DISTINCT(supplierProfileEntity.supplierCategories)')
+//
+//            ->leftJoin(
+//                SupplierProfileEntity::class,
+//                'supplierProfileEntity',
+//                Join::WITH,
+//                'supplierProfileEntity.user = :userId'
+//            )
+//
+//            ->setParameter('userId', $supplierId)
+//
+//            ->getQuery()
+//            ->getSingleColumnResult();
+//
+//        if (empty($query)) {
+//            return $query;
+//        }
+//
+//        return json_decode($query[0]);
+//    }
 }
